% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libFM.R
\name{libFM.model.frame}
\alias{libFM.model.frame}
\title{Converts a data frame into a format digestable for libFM}
\usage{
libFM.model.frame(data, formula)
}
\arguments{
\item{data}{data.frame}

\item{formula}{}
}
\value{
a character vector with one value per observation
}
\description{
This can only deal with factors as covariate.
It is important to make sure your covariates are all factors (not indicators),
or else it will not work correctly
}

