% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_data.R
\name{libFM_groups}
\alias{libFM_groups}
\title{Extract grouping of variables based on formula}
\usage{
libFM_groups(formula, data)
}
\arguments{
\item{formula}{the formula with the response on the left hand side
and the covariates on the right hand side}

\item{data}{data.frame}
}
\value{
vector of integers with group labels of the variables
}
\description{
Extract grouping of variables based on formula
}

