% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manip_data.R
\name{matrix_libFM}
\alias{matrix_libFM}
\title{Converts a matrix to libFM format}
\usage{
matrix_libFM(mat, y)
}
\arguments{
\item{mat}{matrix to be converted. Can either be a standard matrix
or a sparse matrix}

\item{y}{The response variable. If unavailable (for test data), a
  1 is used because libFM requires a response.}
}
\value{
a character vector with one value per observation
}
\description{
Converts a matrix to libFM format
}
\note{
If your data consists of factor variables with a lot of levels,
  model_matrix_libFM is much faster.
}

