% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/manip_data.R
\name{model.frame.libFM}
\alias{model.frame.libFM}
\title{Converts a data frame into a format digestable for libFM}
\usage{
\method{model.frame}{libFM}(formula, data, ...)
}
\arguments{
\item{formula}{the formula with the response on the left hand side
and the covariates on the right hand side}

\item{data}{data.frame}

\item{...}{optional, currently unused, arguments}
}
\value{
a character vector with one value per observation
}
\description{
This can only deal with factors as covariate.
It is important to make sure your covariates are all factors (not indicators),
or else it will not work correctly
}

