% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_data.R
\name{model_frame_libFM}
\alias{model_frame_libFM}
\title{Converts a data frame into a format digestable for libFM}
\usage{
model_frame_libFM(formula, data, ...)
}
\arguments{
\item{formula}{the formula with the response on the left hand side
and the covariates on the right hand side}

\item{data}{data.frame}

\item{...}{optional, currently unused, arguments}
}
\value{
a character vector with one value per observation
}
\description{
This can only deal with factors as covariate.
It is important to make sure your covariates are all factors (not indicators),
or else it will not work correctly
}
\details{
If your data is sparse, \code{sp_matrix_libFM} is about 100 times faster than
  \code{matrix_libFM}. I recommend using the sparse version over the standard version
  whenever possible. If your data consists of factor variables with a lot of levels,
  \code{model_frame_libFM} is faster than \code{sp_matrix_libFM}.
}
\examples{
data(movie_lens)

movie_lens_libFM = model_frame_libFM(Rating ~ User + Movie, movie_lens)
tail(movie_lens_libFM, 10)

}
\seealso{
\code{\link{sp_matrix_libFM}}, \code{\link{matrix_libFM}}
}

