% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_data.R
\name{sp_matrix_libFM}
\alias{sp_matrix_libFM}
\title{Converts a sparse matrix to libFM format}
\usage{
sp_matrix_libFM(mat, y)
}
\arguments{
\item{mat}{sparse matrix to be converted. Must be a sparse matrix
from the package \code{Matrix}}

\item{y}{The response variable. If unavailable (for test data), a
1 is used because libFM requires a response.}
}
\value{
a character vector with one value per observation
}
\description{
Converts a sparse matrix to libFM format
}
\details{
If your data is sparse, sp_matrix_libFM is about 100 times faster than
  matrix_libFM. I recommend using the sparse version over the standard version
  whenever possible. If your data consists of factor variables with a lot of levels,
  model_frame_libFM is faster than sp_matrix_libFM.
}
\examples{
data(movie_lens)
movie_lens_sub = tail(movie_lens, 1000)

# model.matrix will remove the reference level
# which may not be desireable is some situations
movie_lens_mm = Matrix::sparse.model.matrix(Rating ~ User + Movie - 1, data = movie_lens_sub)

movie_lens_libFM = sp_matrix_libFM(movie_lens_mm, movie_lens_sub$Rating)
head(movie_lens_libFM)

}
\seealso{
\code{\link{model_frame_libFM}}, \code{\link{matrix_libFM}}
}

