% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/libFM.R
\name{libFM}
\alias{libFM}
\alias{libFM.data.frame}
\alias{libFM.default}
\alias{libFM.matrix}
\title{libFM factorization machines}
\usage{
libFM(train, test, ...)

\method{libFM}{data.frame}(train, test, formula, validation, ...)

\method{libFM}{matrix}(train, test, y_train, y_test, validation, y_validation,
  ...)

\method{libFM}{default}(train, test, global_bias = TRUE,
  variable_bias = TRUE, dim = 8, task = c("c", "r"), method = c("mcmc",
  "sgd", "als", "sgda"), init_stdev = 0.1, regular = c(0, 0, 0),
  learn_rate = 0.1, validation, verbosity = 0, iter = 100, exe_loc, ...)
}
\arguments{
\item{train}{training data.frame, matrix, or character vector}

\item{test}{testing data.frame, matrix, or character vector}

\item{...}{other, unused, arguments}

\item{formula}{formula of covariates included}

\item{validation}{validation data.frame, matrix, or character vector used for
  adaptive SGD}

\item{y_train,y_test,y_validation}{numeric vectors of responses when train,
  test, and validation are matrices}

\item{global_bias}{whether to include an overall/global bias term}

\item{variable_bias}{whether to include variable main effects/biases}

\item{dim}{dimension of the two-way interaction}

\item{task}{classifcation or regression}

\item{method}{learning method}

\item{init_stdev}{standard deviation used for initialization
of 2-way factors}

\item{regular}{length 3 vector of regularization parameters for
global bias, variable biases, and interactions, respectively. Used with
SGD and ALS}

\item{learn_rate}{learning rate used for SGD and adaptive SGD}

\item{verbosity}{how much feedback to give}

\item{iter}{number of iterations}

\item{exe_loc}{location of libFM.exe executable (if not in the PATH)}
}
\value{
A vector of the predicted values/probabilities
}
\description{
libFM factorization machines
}
\details{
See the libFM manual, \url{http://www.libfm.org/libfm-1.42.manual.pdf},
 for details on the parameters.

 If the function is not working, make sure that the directory is in the PATH
 by running \code{Sys.getenv('PATH')}. It is assumed that the executable is named
 \code{libFM} on. You can verify that the executable is being found and works by
 running \code{system("libFM -help")}.

 See the README on \url{https://github.com/andland/libFMexe} for some more
 information on installation.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: 

\item \code{matrix}: 

\item \code{default}: 
}}
\examples{
\dontrun{
data(movie_lens)
train_rows = sample.int(nrow(movie_lens), nrow(movie_lens) * 2 / 3)
train = movie_lens[train_rows, ]
test  = movie_lens[-train_rows, ]

predFM = libFM(train, test, Rating ~ User + Movie,
               task = "r", dim = 10, iter = 500)

mean((predFM - test$Rating)^2)
}
}
\references{
Steffen Rendle (2012): Factorization Machines with libFM, in ACM Trans.
Intell. Syst. Technol., 3(3), May.
}

