% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip_data.R
\name{matrix_libFM}
\alias{matrix_libFM}
\title{Converts a matrix to libFM format}
\usage{
matrix_libFM(mat, y)
}
\arguments{
\item{mat}{matrix to be converted. Can either be a standard matrix
or a sparse matrix}

\item{y}{The response variable. If unavailable (for test data), a
1 is used because libFM requires a response.}
}
\value{
a character vector with one value per observation
}
\description{
Converts a matrix to libFM format
}
\details{
If your data is sparse, \code{sp_matrix_libFM} is about 100 times faster than
  \code{matrix_libFM}. I recommend using the sparse version over the standard version
  whenever possible. If your data consists of factor variables with a lot of levels,
  \code{model_frame_libFM} is faster than \code{sp_matrix_libFM}.
}
\examples{
data(movie_lens)
movie_lens_sub = tail(movie_lens, 10)

# model.matrix will remove the reference level
# which may not be desireable is some situations
movie_lens_mm = model.matrix(Rating ~ User + Movie - 1, data = movie_lens_sub)

movie_lens_libFM = matrix_libFM(movie_lens_mm, movie_lens_sub$Rating)
movie_lens_libFM

}
\seealso{
\code{\link{sp_matrix_libFM}}, \code{\link{model_frame_libFM}}
}

